+++ src/AiPipeline.TireOcr/AiPipeline.TireOcr.TasyDbMatcher/AiPipeline.TireOcr.TasyDbMatcher.Application/Dtos/TireDbMatch.cs
@@ -3,5 +3,5 @@ namespace AiPipeline.TireOcr.TasyDbMatcher.Application.Dtos;
 public record TireDbMatch(
     ProcessedTireParamsDatabaseEntryDto TireEntry,
     int RequiredCharEdits,
-    double EstimatedAccuracy
+    decimal EstimatedAccuracy
 );
